// src/pages/BackSystem/ArticleManagement/Blog/data.d.ts
export interface aticleDraft {
  articleHtmlContent: string
  articleTabled: string
  articleUrl: string
  createTime: string
  id: number
  labelValues: string
  likes: number
  look: number
  message: string
  name: string
  originalAuthor: string
  selectCategories: string
  selectGrade: number
  selectType: string
  specificTag: string
  tagValue: string[]
  text: string
  title: string
}